10.43.101.191/requirements.txt
mysql-connector-python
requests
pandas
mlflow
scikit-learn
joblib
numpy
sqlalchemy
pymysql
virtualenv
apache-airflow
apache-airflow-providers-postgres
boto3